coreos-koji-tagger: support intermediate tags
Since we'd like the coreos-pool tag (and resulting yum repo)
to be the source of truth for all of our Fedora CoreOS builds
we would like to be able to have different signing keys at play.
We'd like to adopt a policy of getting rpms signed with the key
that matches the buildroot they were targeted against. For example,
if we are including a package from f31 (rawhide) for our next builds
we'd like it to be signed by the f31 key.

In order to achieve this we gather more information about each build
and apply a heuristic function get_releasever_from_buildroottag to
determine which release version of fedora to apply it against. We
then tag it into the intermediate tag that will sign it with that
key before placing it in coreos-pool.