coreos-koji-tagger: move to koji python library
Let's switch to using the koji python library for interfacing
with koji. This is nice so that we don't have to shell out for
each call to koji.

One other noticable thing we do in this commit is add a new
BuildInfo object to hold information about each build that
we use during our processing.

Fixes #2
Obsoletes #28