coreos-koji-tagger: call distrepo manually
We've hit race conditions with the tag2distrepo setup so for now we'll
remove the tag2distrepo config on coreos-pool and we'll just call it
once as part of coreos-koji-tagger after all builds have been tagged
into the coreos-pool tag. This should relieve us of our race condition
for now.

Fixes: https://github.com/coreos/fedora-coreos-releng-automation/issues/52