coreos-ostree-importer: add send-ostree-import-request.py
This script is not used by the importer at all, but rather to send
the request to the importer from the Fedora CoreOS pipeline.

Right now it is configured to just send requests and not wait for
responses. Once we have certificates set up in Fedora infra we'll
modify this to wait for responses before continuing.