fedora-ostree-pruner: generic prune for deleted refs
Just delete the ref first and then we'll do a generic prune for
unreachable objects after looping through all the refs in the repo.

This also means if more than one refs points to a commit we won't delete
it like we were doing with the `ostree prune --delete-commit=$commit`
approach.