coreos-ostree-importer: optimize os.walk
We need to optimize os.walk() here because it can take a really
long time to find all the directories on the NFS mounts because
it must look at every file and there are many many files in the
directories under `objects/*/`. Here we'll optmize for the fact
that we know `objects/*` are directories and `objects/*/*` are normal
files so we don't need to traverse deeper.