coreos-ostree-importer: only do parent commit check on prod repo
Sometimes we do multiple builds in a stream before we release one
(i.e. if we found a problem before release). Each new build will
have a parent commit corresponding to the last *released* build,
which is the correct thing to do. In that case our parent commit
check won't be valid, so let's only do that check on the prod repo.

As part of this I moved the sanity checks out of ostree_pull_local
so I could use the target_repo variable without having to change
the function call signature of ostree_pull_local.