scripts/download-overrides.py: fix regression from archless overrides
The `startswith` check for the archless overrides also matches the
archful ones, which means we were downloading all the RPMs for all
arches.

Really, it's silly to do these loose checks just to support the JSON
version when we don't have any plans to support those (even though cosa
today does support them). Just tighten this all up to check for the
specific filenames we expect.