coreos-koji-tagger: bump signing timeout to 30 mins
We just hit the case where a lot of packages needed to be tagged, and
there was genuine progress being made but the timer ran out. So then I
had to manually run `koji dist-repo`.

Let's just bump the timeout to 30 minutes to catch this.

A smarter method we could follow-up with on this would be to check if
the difference is getting smaller and reset the timeout if so.