fedora-ostree-pruner: generic prune for all unreachable objects
OStree just learned a new `--commit-only` option to `ostree prune` [1].
This allows us to just delete the commits in the history of a ref that
we desire to prune based on our policy, which is a fast operation. After
we've deleted all the commit objects we can then go back and do one
sweep over the repo to now clean up any unreachable objects.

Previously everytime we ran `ostree prune` and deleted a certain amount
of history from a ref the prune operation also computed reachability for
every object in the repo at that time, which is expensive. If you have
many refs this means the entire pruning process can take a long time.
With this new mode of operation deleting the commit objects should take
little time per run and then we can have one big expensive run at the
very end that does the expensive reachability determination for every
object in the repo.