fedora-ostree-pruner: skip over aliased refs
Let's not operate on aliased refs. Rather than require the person
who comes up with the configuration to understand what is an alias
and what isn't let's just do the right thing in the code.

We've seen an issue (we think) where a ref gets deleted and then
running `ostree summary -u` fails and we think it has to do with
operating on aliased refs. Let's find out.