config-bot: support branch-level skip files
With the new experimental stream, we want to have more freedom in what
files go in there without having to also pollute `testing-devel`.

Add support for a `.coreos.skip-files` file which will tell config-bot
to not clobber a specific set of files on that ref. This augments the
global-level `skip-files` list.

Related: https://github.com/coreos/fedora-coreos-tracker/issues/1446