coreos-koji-tagger: tag packages from rawhide builds into pool
For the Bodhi testing work, when testing a rawhide update, we want to
test against a known working rawhide build. To do this, we want to base
our build on the last working rawhide combination. We do this by reusing
the generated lockfile from that build and the git commit of the config
repo.

But for this to work, we need to make sure that we have access to all
the RPMs in that build.

This patch teaches coreos-koji-tagger to listen to the
`coreos.stream.release` fedmsg that the release job emits to know
when a new rawhide build is available; it then downloads the generated
lockfiles and tags in the packages.

Closes: #181